reorganize engine tests and api
