+++ tests/integration/cloudformation/api/test_cloudformation_nested_stacks.py
@@ -13,7 +13,7 @@ def test_nested_stack(s3_client, cfn_client, deploy_cfn_template, s3_create_buck
     s3_client.put_object(
         Bucket=artifacts_bucket,
         Key=artifacts_path,
-        Body=load_file(os.path.join(os.path.dirname(__file__), "../templates/template5.yaml")),
+        Body=load_file(os.path.join(os.path.dirname(__file__), "../../templates/template5.yaml")),
     )
 
     # deploy template
@@ -21,7 +21,9 @@ def test_nested_stack(s3_client, cfn_client, deploy_cfn_template, s3_create_buck
     stack_bucket_name = f"test-{param_value}"  # this is the bucket name generated by template5
 
     deploy_cfn_template(
-        template=load_file(os.path.join(os.path.dirname(__file__), "../templates/template6.yaml"))
+        template=load_file(
+            os.path.join(os.path.dirname(__file__), "../../templates/template6.yaml")
+        )
         % (artifacts_bucket, artifacts_path),
         parameters={"GlobalParam": param_value},
     )
@@ -41,14 +43,14 @@ def test_nested_stack_output_refs(cfn_client, deploy_cfn_template, s3_client, s3
     key = f"test-key-{short_uid()}"
     s3_client.upload_file(
         os.path.join(
-            os.path.dirname(__file__), "../templates/nested-stack-output-refs.nested.yaml"
+            os.path.dirname(__file__), "../../templates/nested-stack-output-refs.nested.yaml"
         ),
         Bucket=bucket_name,
         Key=key,
     )
     result = deploy_cfn_template(
         template_path=os.path.join(
-            os.path.dirname(__file__), "../templates/nested-stack-output-refs.yaml"
+            os.path.dirname(__file__), "../../templates/nested-stack-output-refs.yaml"
         ),
         template_mapping={
             "s3_bucket_url": f"/{bucket_name}/{key}",
