+++ tests/integration/cloudformation/api/test_cloudformation_stacks.snapshot.json
