+++ tests/integration/cloudformation/test_cloudformation_engine.snapshot.json
@@ -1,5 +1,5 @@
 {
-  "tests/integration/cloudformation/dynamic_template/test_cloudformation_engine.py::test_implicit_type_conversion": {
+  "tests/integration/cloudformation/test_cloudformation_engine.py::TestTypes::test_implicit_type_conversion": {
     "recorded-date": "12-08-2022, 16:14:04",
     "recorded-content": {
       "queue": {
